API: Move semantic types into `marker_api::sem::ty`
