+++ marker_api/src/ast/ty/other_ty.rs
@@ -1,21 +1,4 @@
-use std::marker::PhantomData;
-
 use super::CommonSynTyData;
-
-/// The placeholder type, signalling that the semantic type is still unstable
-/// and therefor not represented as part of the API.
-#[repr(C)]
-#[derive(Debug)]
-pub struct SemUnstableTy<'ast> {
-    _lt: PhantomData<&'ast ()>,
-}
-
-#[cfg(feature = "driver-api")]
-impl<'ast> SemUnstableTy<'ast> {
-    pub fn new() -> Self {
-        Self { _lt: PhantomData }
-    }
-}
 
 #[repr(C)]
 #[derive(Debug)]
