+++ marker_api/src/ast/ty/sequence_ty.rs
@@ -1,9 +1,9 @@
 use crate::{
-    ast::{expr::ConstExpr, ConstValue},
+    ast::expr::ConstExpr,
     ffi::{FfiOption, FfiSlice},
 };
 
-use super::{CommonSynTyData, SemTyKind, SynTyKind};
+use super::{CommonSynTyData, SynTyKind};
 
 /// The syntactic representation of a tuple type like [`()`](prim@tuple) or [`(T, U)`](prim@tuple)
 #[repr(C)]
@@ -32,38 +32,6 @@ impl<'ast> SynTupleTy<'ast> {
 }
 
 impl<'ast> std::fmt::Display for SynTupleTy<'ast> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut f = f.debug_tuple("");
-
-        for entry in self.types.as_slice() {
-            f.field(entry);
-        }
-
-        f.finish()
-    }
-}
-
-/// The semantic representation of a tuple type like [`()`](prim@tuple) or [`(T, U)`](prim@tuple)
-#[repr(C)]
-#[derive(Debug)]
-pub struct SemTupleTy<'ast> {
-    types: FfiSlice<'ast, SemTyKind<'ast>>,
-}
-
-impl<'ast> SemTupleTy<'ast> {
-    pub fn types(&self) -> &[SemTyKind<'ast>] {
-        self.types.as_slice()
-    }
-}
-
-#[cfg(feature = "driver-api")]
-impl<'ast> SemTupleTy<'ast> {
-    pub fn new(types: &'ast [SemTyKind<'ast>]) -> Self {
-        Self { types: types.into() }
-    }
-}
-
-impl<'ast> std::fmt::Display for SemTupleTy<'ast> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let mut f = f.debug_tuple("");
 
@@ -104,31 +72,6 @@ impl<'ast> std::fmt::Display for SynSliceTy<'ast> {
     }
 }
 
-/// The semantic representation of a variable length slice like [`[T]`](prim@slice)
-#[repr(C)]
-pub struct SemSliceTy<'ast> {
-    inner_ty: SemTyKind<'ast>,
-}
-
-impl<'ast> SemSliceTy<'ast> {
-    pub fn inner_ty(&self) -> SemTyKind<'ast> {
-        self.inner_ty
-    }
-}
-
-#[cfg(feature = "driver-api")]
-impl<'ast> SemSliceTy<'ast> {
-    pub fn new(inner_ty: SemTyKind<'ast>) -> Self {
-        Self { inner_ty }
-    }
-}
-
-impl<'ast> std::fmt::Debug for SemSliceTy<'ast> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_list().entries(std::iter::once(self.inner_ty())).finish()
-    }
-}
-
 /// The syntactic representation of an array with a known size like: [`[T; N]`](prim@array)
 #[repr(C)]
 #[derive(Debug)]
@@ -164,38 +107,6 @@ impl<'ast> SynArrayTy<'ast> {
 }
 
 impl<'ast> std::fmt::Display for SynArrayTy<'ast> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // FIXME: Add length expression
-        f.debug_list().entries(std::iter::once(self.inner_ty())).finish()
-    }
-}
-
-/// The semantic representation of an array with a known size like: [`[T; N]`](prim@array)
-#[repr(C)]
-#[derive(Debug)]
-pub struct SemArrayTy<'ast> {
-    inner_ty: SemTyKind<'ast>,
-    len: ConstValue<'ast>,
-}
-
-impl<'ast> SemArrayTy<'ast> {
-    pub fn inner_ty(&self) -> SemTyKind<'ast> {
-        self.inner_ty
-    }
-
-    pub fn len(&self) -> &ConstValue<'ast> {
-        &self.len
-    }
-}
-
-#[cfg(feature = "driver-api")]
-impl<'ast> SemArrayTy<'ast> {
-    pub fn new(inner_ty: SemTyKind<'ast>, len: ConstValue<'ast>) -> Self {
-        Self { inner_ty, len }
-    }
-}
-
-impl<'ast> std::fmt::Display for SemArrayTy<'ast> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         // FIXME: Add length expression
         f.debug_list().entries(std::iter::once(self.inner_ty())).finish()
