+++ marker_api/src/lib.rs
@@ -11,11 +11,13 @@
 pub static MARKER_API_VERSION: &str = env!("CARGO_PKG_VERSION");
 
 pub mod ast;
+pub mod common;
 pub mod context;
 pub mod diagnostic;
 pub mod interface;
 pub mod lint;
 pub mod prelude;
+pub mod sem;
 pub mod span;
 
 #[doc(hidden)]
