+++ marker_api/src/prelude.rs
@@ -21,9 +21,9 @@ pub use crate::ast::item::Body;
 pub use crate::ast::item::ItemKind;
 pub use crate::ast::pat::PatKind;
 pub use crate::ast::stmt::StmtKind;
-pub use crate::ast::ty::SemTyKind;
 pub use crate::ast::ty::SynTyKind;
 pub use crate::lint::Lint;
+pub use crate::sem::ty::SemTyKind;
 pub use crate::span::Ident;
 pub use crate::span::Span;
 pub use crate::MarkerContext;
