+++ marker_api/src/sem.rs
@@ -1,0 +1,7 @@
+//! A module containing semantic representations of types, generics and other
+//! parts of the language.
+
+mod common;
+pub use common::*;
+
+pub mod ty;
