+++ marker_rustc_driver/src/conversion/marker/generics.rs
@@ -1,11 +1,14 @@
-use marker_api::ast::{
-    generic::{
-        Lifetime, LifetimeKind, SemBindingArg, SemConstArg, SemGenericArgKind, SemGenericArgs, SemTraitBound,
-        SynBindingArg, SynConstArg, SynConstParam, SynGenericArgKind, SynGenericArgs, SynGenericParamKind,
-        SynGenericParams, SynLifetimeArg, SynLifetimeClause, SynLifetimeParam, SynTraitBound, SynTyArg, SynTyClause,
-        SynTyParam, SynTyParamBound, SynWhereClauseKind,
+use marker_api::{
+    ast::{
+        generic::{
+            Lifetime, LifetimeKind, SemBindingArg, SemConstArg, SemGenericArgKind, SemGenericArgs, SemTraitBound,
+            SynBindingArg, SynConstArg, SynConstParam, SynGenericArgKind, SynGenericArgs, SynGenericParamKind,
+            SynGenericParams, SynLifetimeArg, SynLifetimeClause, SynLifetimeParam, SynTraitBound, SynTyArg,
+            SynTyClause, SynTyParam, SynTyParamBound, SynWhereClauseKind,
+        },
+        TraitRef,
     },
-    ConstValue, TraitRef,
+    sem::ConstValue,
 };
 use rustc_hir as hir;
 use rustc_middle as mid;
